I Did It
I see what is on the other side of the hill...

Anyways the offending problem was ActorFactory.hpp (Teacher Provided -.-)  and that the switch statement didn't have a default case. Or at least that is what did it for me. 
I also fixed some loop errors in Battle.cpp; it had the false assignment outside the loop which didn't help. Really I only worked on Battle.cpp and Actor.hpp and that seemed to get everything in working order.
Everything non-intuitive is commented, I think.
Now I have to read four physics chapters in one day and one night...